third_party/blink/web_tests/http/tests/security/no-indexeddb-from-sandbox-expected.txt

CONSOLE MESSAGE: PASS: indexedDB.open() threw a SECURITY_ERR!
